1、优化凭证过期时间bug 2、expiredSeconds应该用long类型存储,如果用int类型,在换算成毫秒时会超出int的范围,造成数据丢失

+++ src/main/java/com/my/community/service/IUserService.java
@@ -19,7 +19,7 @@ public interface IUserService {
 
     Map<String, Object> register(User user);
 
-    Map<String, Object> login(String username, String password, int expiredSeconds);
+    Map<String, Object> login(String username, String password, long expiredSeconds);
 
     void logout(String ticket);
